@@ -2,35 +2,37 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 
 class Rounds{
-  final int round;
-  final int table;
-  final int board;
-  final String teamNS;
-  final String teamEW;
-  int teamNSScore;
-  int teamEWScore;
+  final String roundID;
+  final int round_NO;
+  final String table_NO;
+  final List board_NO;
+  final String team_NS;
+  final String team_EW;
+  // int declarer;
+  // int score;
 
   Rounds({
-    @required this.round,
-    @required this.table,
-    @required this.board,
-    @required this.teamNS,
-    @required this.teamEW,
+    @required this.roundID,
+    @required this.round_NO,
+    @required this.table_NO,
+    @required this.board_NO,
+    @required this.team_NS,
+    @required this.team_EW,
   });
 
-  int get teamNSScores{
-    return teamNSScore;
-  }
+  // int get teamNSScores{
+  //   return teamNSScore;
+  // }
 
-  set teamNSScores(int teamNSScore){
-    this.teamNSScore = teamNSScore;
-  }
+  // set teamNSScores(int teamNSScore){
+  //   this.teamNSScore = teamNSScore;
+  // }
 
-  int get teamEWScores{
-    return teamNSScore;
-  }
+  // int get teamEWScores{
+  //   return teamNSScore;
+  // }
 
-  set teamEWScores(int teamEWScore){
-    this.teamEWScore = teamEWScore;
-  }
+  // set teamEWScores(int teamEWScore){
+  //   this.teamEWScore = teamEWScore;
+  // }
 }
