@@ -1,7 +1,4 @@
-import math
-
-trump = [9, 5, 1]
-# trump = [7, 7, 4]
+trump = [7, 6, 4]
 
 print('\tNT\tMJ\tMN')
 print("--------------------------------")
@@ -10,8 +7,8 @@ for i in range(7):
     print(level, end='   |\t')
     for j in range(len(trump)):
         weight = int(
-            # math.pow(2, level) - level + trump[j]
-            math.pow(2, level) - level + trump[j]
+
+            2*level + trump[j]
         )
         print(weight, end='\t')
     print()
