@@ -1,24 +1,29 @@
 import 'package:flutter/material.dart';
-// import '../models/calculating.dart';
+import '../../methods/calculator.dart';
 
 class ContractEditor extends StatefulWidget {
+  final Calculator recentCalculator;
+  ContractEditor(this.recentCalculator);
+
   @override
   _ContractEditorState createState() => _ContractEditorState();
 }
 
 class _ContractEditorState extends State<ContractEditor> {
-  int contractPoint = 1;
+  int _contractPoint = 1;
 
 
-  void _contractIncrese() {
+  void contractIncrese() {
     setState(() {
-      contractPoint += 1;
+      _contractPoint += 1;
+      widget.recentCalculator.contractValues = _contractPoint;
     });
   }
 
-  void _contractDecrese() {
+  void contractDecrese() {
     setState(() {
-      contractPoint -= 1;
+      _contractPoint -= 1;
+      widget.recentCalculator.contractValues = _contractPoint;
     });
   }
 
@@ -34,7 +39,7 @@ class _ContractEditorState extends State<ContractEditor> {
             child: Card(
               child: Center(
                 child: const Text(
-                  'contract',
+                  'Contract',
                   style: TextStyle(
                     fontSize: 20,
                     color: Colors.black87,
@@ -54,10 +59,10 @@ class _ContractEditorState extends State<ContractEditor> {
                     leading: IconButton(
                       icon: Icon(Icons.arrow_back_ios),
                       color: Theme.of(context).errorColor,
-                      onPressed: contractPoint <= 1 ? null : _contractDecrese,
+                      onPressed: _contractPoint <= 1 ? null : contractDecrese,
                     ),
                     title: Text(
-                      '$contractPoint',
+                      '$_contractPoint',
                       style: TextStyle(
                         fontWeight: FontWeight.bold,
                         fontSize: 30,
@@ -68,7 +73,7 @@ class _ContractEditorState extends State<ContractEditor> {
                     trailing: IconButton(
                       icon: Icon(Icons.arrow_forward_ios),
                       color: Theme.of(context).errorColor,
-                      onPressed: contractPoint >= 7 ? null : _contractIncrese,
+                      onPressed: _contractPoint >= 7 ? null : contractIncrese,
                     ),
                   ),
                 ),
