@@ -1,19 +1,21 @@
 import 'package:flutter/material.dart';
-//import '../display/made_display.dart';
+import '../../methods/calculator.dart';
 
 class MadeEditor extends StatefulWidget {
+  final Calculator recentCalculator;
+  MadeEditor(this.recentCalculator);
+
   @override
   _MadeEditorState createState() => _MadeEditorState();
 }
 
 class _MadeEditorState extends State<MadeEditor> {
   int _made = 7;
-  // final MadeDisplay madeDisplay;
 
   void _madeIncrese() {
     setState(() {
       _made += 1;
-      
+      widget.recentCalculator.madeValues = _made;
     });
     
   }
@@ -21,9 +23,13 @@ class _MadeEditorState extends State<MadeEditor> {
   void _madeDecrese() {
     setState(() {
       _made -= 1;
-      // MadeDisplay(_made);
+      widget.recentCalculator.madeValues = _made;
     });
     
+  }
+
+  int get thisMade {
+    return _made;
   }
 
   @override
@@ -36,15 +42,14 @@ class _MadeEditorState extends State<MadeEditor> {
             width: 100,
             height: 60,
             child: Card(
-              color: Colors.black54,
+              color: Colors.black38,
               child: Center(
                 child: const Text(
-                  'made',
+                  'Made',
                   style: TextStyle(
                     fontSize: 20,
                     color: Colors.white,
                   ),
-                  
                 ),
               ),
             ),
@@ -54,7 +59,7 @@ class _MadeEditorState extends State<MadeEditor> {
             child: Container(
               height: 60,
               child: Card(
-                color: Colors.black54,
+                color: Colors.black38,
                 child: ListTile(
                   leading: IconButton(
                     icon: Icon(Icons.arrow_back_ios),
